src/main/java/com/packt/app/album/AlbumServiceImpl.java
package com.packt.app.album;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class AlbumServiceImpl implements AlbumService{

    private AlbumRepository albumRepository;

    @Autowired
    public AlbumServiceImpl(AlbumRepository albumRepository) {
        this.albumRepository = albumRepository;
    }

    public List<Album> getAlbums(){
        return (List<Album>) albumRepository.findAll();
    }
    public void saveAlbums(Album album){
        albumRepository.save(album);
    }
}
